Create menu.java
The Blackjack game doesn't have a method call to play the whole game, and there is not code for anything else but once you get it you can call the methods in the switch statements. Also when I close the ATM when calling it through menu it doesn't return back to menu which idk how to fix.